Refactor core.payments file structure and related imports and tests
