+++ concent_api/core/tests/test_auth_accept_or_reject_results.py
@@ -65,7 +65,7 @@ class AuthAcceptOrRejectIntegrationTest(ConcentIntegrationTestCase):
         )
 
         with mock.patch(
-            'core.message_handlers.core.payments.base.is_account_status_positive',
+            'core.message_handlers.payments_service.is_account_status_positive',
             side_effect=self.is_account_status_positive_true_mock
         ) as is_account_status_positive_true_mock_function:
             with freeze_time("2018-02-05 10:00:30"):
@@ -215,7 +215,7 @@ class AuthAcceptOrRejectIntegrationTest(ConcentIntegrationTestCase):
         )
 
         with mock.patch(
-            'core.message_handlers.core.payments.base.is_account_status_positive',
+            'core.message_handlers.payments_service.is_account_status_positive',
             side_effect=self.is_account_status_positive_true_mock
         ) as is_account_status_positive_true_mock_function:
             with freeze_time("2018-02-05 10:00:31"):
@@ -342,7 +342,7 @@ class AuthAcceptOrRejectIntegrationTest(ConcentIntegrationTestCase):
         )
 
         with mock.patch(
-            'core.message_handlers.core.payments.base.is_account_status_positive',
+            'core.message_handlers.payments_service.is_account_status_positive',
             side_effect=self.is_account_status_positive_true_mock
         ) as is_account_status_positive_true_mock_function:
             with freeze_time("2018-02-05 10:00:30"):
@@ -677,7 +677,7 @@ class AuthAcceptOrRejectIntegrationTest(ConcentIntegrationTestCase):
         )
 
         with mock.patch(
-            'core.message_handlers.core.payments.base.is_account_status_positive',
+            'core.message_handlers.payments_service.is_account_status_positive',
             side_effect=self.is_account_status_positive_true_mock
         ) as is_account_status_positive_true_mock_function:
             with freeze_time("2018-02-05 10:00:30"):
@@ -1019,7 +1019,7 @@ class AuthAcceptOrRejectIntegrationTest(ConcentIntegrationTestCase):
         )
 
         with mock.patch(
-            'core.message_handlers.core.payments.base.is_account_status_positive',
+            'core.message_handlers.payments_service.is_account_status_positive',
             side_effect=self.is_account_status_positive_true_mock
         ) as is_account_status_positive_true_mock_function:
             with freeze_time("2018-02-05 10:00:30"):
@@ -1222,7 +1222,7 @@ class AuthAcceptOrRejectIntegrationTest(ConcentIntegrationTestCase):
         )
 
         with mock.patch(
-            'core.message_handlers.core.payments.base.is_account_status_positive',
+            'core.message_handlers.payments_service.is_account_status_positive',
             side_effect=self.is_account_status_positive_true_mock
         ) as is_account_status_positive_true_mock_function:
             with freeze_time("2018-02-05 10:00:30"):
