+++ concent_api/core/tests/test_auth_force_payment.py
@@ -57,7 +57,7 @@ class AuthForcePaymentIntegrationTest(ConcentIntegrationTestCase):
         )
 
         with mock.patch(
-            'core.message_handlers.core.payments.base.make_force_payment_to_provider',
+            'core.message_handlers.payments_service.make_force_payment_to_provider',
             side_effect=self._make_force_payment_to_provider
         ) as make_force_payment_to_provider_mock:
             with freeze_time("2018-02-05 12:00:20"):
@@ -66,7 +66,7 @@ class AuthForcePaymentIntegrationTest(ConcentIntegrationTestCase):
                     self._get_list_of_force_transactions(),
                 ]
                 with mock.patch(
-                    'core.message_handlers.core.payments.base.get_list_of_payments',
+                    'core.message_handlers.payments_service.get_list_of_payments',
                     side_effect=fake_responses
                 ) as get_list_of_payments_mock:
                     response_1 = self.client.post(
